fixes 7 failing pattern matching tests due to an incorrect assumption in the implementation of EnsureEnglishCulture, namely that the English culture always uses standard cultural settings; the fix now always ensures the use of the invariant culture
